Add progress and bug fixes
