+++ .vscode/yaml.code-snippets
@@ -1,0 +1,41 @@
+{
+	// Place your snippets for yaml here. Each snippet is defined under a snippet name and has a prefix, body and 
+	// description. The prefix is what is used to trigger the snippet and the body will be expanded and inserted. Possible variables are:
+	// $1, $2 for tab stops, $0 for the final cursor position, and ${1:label}, ${2:another} for placeholders. Placeholders with the 
+	// same ids are connected.
+	// Example:
+	// "Print to console": {
+	// 	"prefix": "log",
+	// 	"body": [
+	// 		"console.log('$1');",
+	// 		"$2"
+	// 	],
+	// 	"description": "Log output to console"
+	// }
+	"Insert OpenFisca basic parameter": {
+		"prefix": "param",
+		"body": [
+			"description: ${1:Description}",
+			"values:",
+			"  2019-06-01:",
+			"    value: ${2:0.00}",
+			"metadata:",
+			"  unit: currency-GBP",
+			"  reference: $3:none"
+		],
+		"description": "Insert a basic OpenFisca parameter."
+	},
+	"Insert OpenFisca basic test": {
+		"prefix": "test",
+		"body": [
+			"- name: ${1:Name}",
+			"  period: 2020",
+			"  absolute_error_margin: ${2:0}",
+			"  input:",
+			"    ${3:variable values here}",
+			"  output:",
+			"    ${4:expected output here}"
+		],
+		"description": "Insert a basic OpenFisca parameter for one person."
+	}
+}
